feature #1 - Add donut object to the scene, with camera controls. Improve README (#2)
* feat: display donut in the scene

I've update the README to better reflect the project, which now also includes a section for
development. I've added a 3D donut to the scene. Lastly, I've modularized the code to make it less
coupled.

* feat: make donut bop, introduce camera controls

Added the ability to room in and out, as well as the ability to rotate the camera. Re-colored the
donut. Made the donut bop up and down.

fix #1